Remove Quorum
