--- plaid/resources/Cargo.toml
@@ -1,71 +1,0 @@
-[package]
-name = "plaid"
-version = "0.14.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-default = ["aws"]
-quorum = ["quorum-agent"]
-aws = ["aws-sdk-kms", "aws-config"]
-
-[dependencies]
-alkali = "0.3.0"
-async-trait = "0.1.56"
-base64 = "0.13"
-clap = { version = "4", default-features = false, features = ["std"] }
-crossbeam-channel = "0.5"
-env_logger = "0.8"
-flate2 = "1.0"
-hex = "0.4.3"
-http = "1"
-jwt-simple = "0.12.10"
-log = "0.4"
-lru = "0.12"
-octocrab = "0.37"
-paste = "1.0"
-plaid_stl = { path = "../plaid-stl" }
-rcgen = { version = "0.10", features = ["x509-parser"] }
-regex = "1"
-ring = "0.17"
-reqwest = { version = "0.11", default-features = false, features = [
-    "rustls-tls",
-    "json",
-    "cookies",
-] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1"
-sled = "0.34.7"
-tar = "0.4.41"
-time = "0.3"
-tokio = { version = "1", features = ["full"] }
-toml = "0.5"
-totp-rs = "5.6.0"
-url = "2.5.2"
-urlencoding = "2.1.3"
-warp = { version = "0.3", features = ["tls"] }
-wasmer = { version = "4", default-features = false, features = ["cranelift"] }
-wasmer-middlewares = "4"
-jsonwebtoken = { version = "9.2" }
-tokio-tungstenite = { version = "0.23.1", features = ["native-tls-vendored"] }
-futures-util = "0.3.30"
-aws-sdk-kms = { version = "1.41.0", optional = true }
-aws-config = { version = "1.5.5", optional = true }
-
-# Uncomment to build with Quorum. This is needed
-# because otherwise cargo will try and find this
-# to build the lockfile.
-quorum-agent = { path = "../../quorum/quorum-agent", default_features = false, optional = true }
-
-[[example]]
-name = "github-tailer"
-path = "examples/tailers/github.rs"
-
-[[bin]]
-name = "plaid"
-path = "src/bin/plaid.rs"
-
-[[bin]]
-name = "config_check"
-path = "src/bin/config_check.rs"
