--- plaid/resources/Dockerfile_quorum.aarch64
@@ -1,22 +1,0 @@
-FROM messense/rust-musl-cross:aarch64-musl as builder
-
-RUN rustup component add rustfmt
-RUN mkdir /build
-WORKDIR /build
-COPY plaid plaid
-COPY quorum quorum
-
-COPY plaid/plaid/resources/Cargo.toml plaid/plaid/Cargo.toml
-
-WORKDIR /build/plaid
-RUN cargo build --bin=plaid --release --features quorum
-
-
-FROM alpine:3.6 as alpine
-RUN apk add -U --no-cache ca-certificates
-
-from scratch as runtime
-COPY --from=alpine /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
-COPY --from=builder /build/plaid/target/aarch64-unknown-linux-musl/release/plaid /plaid
-USER 1000
-CMD [ "/plaid", "--config", "/config/plaid.toml", "--secrets", "/config/secrets.json" ]
