+++ plaid/src/functions/api.rs
@@ -279,9 +279,6 @@ impl_new_function_with_error_buffer!(okta, get_user_data);
 // PagerDuty Functions
 impl_new_function!(pagerduty, trigger_incident);
 
-// Quorum Functions
-impl_new_function_with_error_buffer!(quorum, proposal_status);
-
 // Rustica Functions
 impl_new_function_with_error_buffer!(rustica, new_mtls_cert);
 
@@ -336,16 +333,16 @@ pub fn to_api_function(
         }
         "get_time" => Function::new_typed(&mut store, super::internal::get_time),
         "storage_insert" => {
-            Function::new_typed_with_env(&mut store, &env, super::internal::storage_insert)
+            Function::new_typed_with_env(&mut store, &env, super::storage::storage_insert)
         }
         "storage_get" => {
-            Function::new_typed_with_env(&mut store, &env, super::internal::storage_get)
+            Function::new_typed_with_env(&mut store, &env, super::storage::storage_get)
         }
 
         "cache_insert" => {
-            Function::new_typed_with_env(&mut store, &env, super::internal::cache_insert)
+            Function::new_typed_with_env(&mut store, &env, super::cache::cache_insert)
         }
-        "cache_get" => Function::new_typed_with_env(&mut store, &env, super::internal::cache_get),
+        "cache_get" => Function::new_typed_with_env(&mut store, &env, super::cache::cache_get),
         "log_back" => Function::new_typed_with_env(&mut store, &env, super::internal::log_back),
         "log_back_unlimited" => Function::new_typed_with_env(&mut store, &env, super::internal::log_back_unlimited),
         
@@ -513,11 +510,6 @@ pub fn to_api_function(
         // PagerDuty Calls
         "pagerduty_trigger_incident" => {
             Function::new_typed_with_env(&mut store, &env, pagerduty_trigger_incident)
-        }
-
-        // Quorum Calls
-        "quorum_proposal_status" => {
-            Function::new_typed_with_env(&mut store, &env, quorum_proposal_status)
         }
 
         // Rustica Calls
