+++ plaid/src/functions/mod.rs
@@ -1,7 +1,9 @@
 mod api;
+mod cache;
 mod internal;
 mod memory;
 mod message;
+mod storage;
 
 use memory::*;
