+++ plaid/src/storage/sled/mod.rs
@@ -75,16 +75,20 @@ impl StorageProvider for Sled {
         Ok(result.map(|v| v.to_vec()))
     }
 
-    async fn list_keys(&self, namespace: &str) -> Result<Vec<Vec<u8>>, StorageError> {
+    async fn list_keys(&self, namespace: &str, prefix: Option<&str>) -> Result<Vec<Vec<u8>>, StorageError> {
         let tree = self
             .db
             .open_tree(namespace.as_bytes())
             .map_err(|_| StorageError::Access(format!("Could not open Sled tree {namespace}")))?;
 
+
+        let key_iter = match prefix {
+            Some(p) => tree.scan_prefix(p),
+            None => tree.iter()
+        };
         // The use of a filter_map here means keys that fail to be pulled will be thrown away.
         // I don't know if this is possible? Maybe if the database is moved out from under us?
-        let keys: Vec<Vec<u8>> = tree
-            .iter()
+        let keys: Vec<Vec<u8>> = key_iter
             .keys()
             .filter_map(|x| match x {
                 Ok(v) => Some(v.to_vec()),
@@ -98,16 +102,19 @@ impl StorageProvider for Sled {
         Ok(keys)
     }
 
-    async fn fetch_all(&self, namespace: &str) -> Result<Vec<(Vec<u8>, Vec<u8>)>, StorageError> {
+    async fn fetch_all(&self, namespace: &str, prefix: Option<&str>) -> Result<Vec<(Vec<u8>, Vec<u8>)>, StorageError> {
         let tree = self
             .db
             .open_tree(namespace.as_bytes())
             .map_err(|_| StorageError::Access(format!("Could not open Sled tree {namespace}")))?;
 
+        let key_iter = match prefix {
+            Some(p) => tree.scan_prefix(p),
+            None => tree.iter()
+        };
         // The use of a filter_map here means keys that fail to be pulled will be thrown away.
         // I don't know if this is possible? Maybe if the database is moved out from under us?
-        let data: Vec<(Vec<u8>, Vec<u8>)> = tree
-            .iter()
+        let data: Vec<(Vec<u8>, Vec<u8>)> = key_iter
             .filter_map(|x| match x {
                 Ok((k, v)) => Some((k.to_vec(), v.to_vec())),
                 Err(e) => {
